requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
kaleido>=0.2.1  # For static image export in Plotly

# Machine Learning (for anomaly detection and feature engineering)
scikit-learn>=1.3.0

# File format support
openpyxl>=3.1.0
xlrd>=2.0.1

# Streamlit UI Dashboard
streamlit>=1.28.0
streamlit-plotly-events>=0.0.6
watchdog>=3.0.0  # For Streamlit file watching

# Optional but useful
jupyter>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
